Implement generateRandomV2landString
