+++ api/services/UtilService.js
@@ -24,6 +24,20 @@ module.exports = {
     return newFilename;
   },
 
+  generateRandomV2landString: length => {
+    const charset = ['🧜‍♀️', '🧜‍♂️', '🐠', '🐟', '🐬', '🐳', '🐋',
+      '💧', '💦', '🌊', '🏄‍♀️', '🏄‍♂️', '🏊‍♀️', '🏊‍♂️', '🚣‍♀️', '🚣‍♂️',
+      '🚤', '🛥', '⛵️', '🛶', '🛳', '⛴', '🚢', '💙', '🏖', '🏝'];
+
+    let string = '';
+    for (let i = 0; i < length; i++) {
+      const index = Math.floor(Math.random() * charset.length);
+      string += charset[index];
+    }
+
+    return string;
+  },
+
   /**
    * 生成 WHERE 语句
    */
